requirements.txt
streamlit
langchain-openai
openai
wikipedia
python-dotenv
requests
